basis functions mostly ported
